Implement top-level metadata for page data
This commit adds the capability of associating arbitrary metadata with
the pages in the rendering context by lifting keys and values out of
forms that have metadata with the :page prefix keyword into the
top-level metadata var, and then lifting that data out of the
metadata var into the page map.

It also moves away from the practice of storing full (ns) forms as
metadata; either a page has an actual namespace declaration inline, or
it names its namespace as a symbol.