Begin extending Flexmark to support WikiLinks in Cybermonday Hiccup conversion
Right now, there's no attempt at link resolution. This might be OK because the
contents of the links are just data and can be resolved/postprocessed after the
parsing step. However, this implementation does not yet support the :lower-fns
that allow for flexible transformation from the new IR field of
:markdown/wikilink to standard Hiccup :a link elements.